requirements.txt
flask==2.0.1
flask-login==0.5.0
mysqlclient==2.0.3
python-dotenv==0.19.0
scikitlearn==0.24.2
nltk==3.6.2
pandas==1.3.2
numpy==1.21.2
requests==2.26.0